+++ test/dummy/config/importmap.rb
@@ -4,6 +4,7 @@ pin "application"
 pin "@hotwired/stimulus", to: "stimulus.min.js"
 pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
 pin_all_from "app/javascript/controllers", under: "controllers"
+pin_all_from RecordingStudioTermsAndConditions::Engine.root.join("app/javascript/recording_studio_terms_and_conditions/controllers"), under: "controllers/recording_studio_terms_and_conditions", to: "recording_studio_terms_and_conditions/controllers", preload: false
 
 # Pin FlatPack controllers
 pin_all_from FlatPack::Engine.root.join("app/javascript/flat_pack/controllers"), under: "controllers/flat_pack", to: "flat_pack/controllers", preload: false